feat: new routing with redirects for now
