+++ src/pages/api/listings/submissions/[slug].ts
@@ -25,6 +25,7 @@ export default async function user(req: NextApiRequest, res: NextApiResponse) {
             name: true,
             logo: true,
             isVerified: true,
+            slug: true,
           },
         },
         poc: {
